[SG-70] modify: move approot vm in vm module
