+++ app/src/main/kotlin/com/captures2024/soongan/route/AppRoute.kt
@@ -3,8 +3,8 @@ package com.captures2024.soongan.route
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import com.captures2024.soongan.AppRootViewModel
 import com.captures2024.soongan.core.designsystem.component.SoonGanBackground
+import com.captures2024.soongan.core.viewmodel.AppRootViewModel
 import com.captures2024.soongan.feature.intro.route.IntroRoute
 import com.captures2024.soongan.feature.main.route.MainRoute
 import com.captures2024.soongan.feature.sign.route.SignRoute
