--- app/src/main/kotlin/com/captures2024/soongan/state/AppRootRouteState.kt
@@ -1,7 +1,0 @@
-package com.captures2024.soongan.state
-
-enum class AppRootRouteState {
-    LANDING,
-    SIGN,
-    MAIN
-}
