+++ core/viewmodel/build.gradle.kts
@@ -12,6 +12,7 @@ android {
 
 dependencies {
     implementation(projects.core.analytics)
+    implementation(projects.core.common)
     implementation(projects.core.domain)
     implementation(projects.core.model)
