+++ core/viewmodel/src/main/kotlin/com/captures2024/soongan/core/viewmodel/state/AppRootUIState.kt
@@ -1,12 +1,13 @@
-package com.captures2024.soongan.state
+package com.captures2024.soongan.core.viewmodel.state
 
 import com.captures2024.soongan.core.analytics.utils.LogElementArgument
 import com.captures2024.soongan.core.common.base.UIState
 import com.captures2024.soongan.core.model.dto.UserInfoDto
+import com.captures2024.soongan.core.viewmodel.utils.AppRootRoute
 
-internal data class AppRootUIState(
+data class AppRootUIState(
     val isLoading: Boolean = false,
-    val rootRouteState: AppRootRouteState = AppRootRouteState.LANDING,
+    val rootRouteState: AppRootRoute = AppRootRoute.LANDING,
     private val memberInfo: UserInfoDto = UserInfoDto.defaultBuilder(),
 ) : UIState {
