+++ core/viewmodel/src/main/kotlin/com/captures2024/soongan/core/viewmodel/utils/AppRootRoute.kt
@@ -1,0 +1,7 @@
+package com.captures2024.soongan.core.viewmodel.utils
+
+enum class AppRootRoute {
+    LANDING,
+    SIGN,
+    MAIN
+}
